requirements.txt
requests
matplotlib
python-dotenv
streamlit
plotly
flask
